requirements.txt
beautifulsoup4==4.13.4
blinker==1.8.2
cement==2.6.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.4
droopescan==1.45.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==2.1.5
pycparser==2.22
pyOpenSSL==25.1.0
pystache==0.6.8
requests==2.32.4
soupsieve==2.7
SQLAlchemy==2.0.41
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
